2021/12/26ゲーム機能移植中
プライヤー移動まで
ライトの位置が微妙

+++ FrauEngine/FrauEngine/Src/input.h
@@ -65,8 +65,8 @@ public:
 		return &mouseInput;
 	}
 public:
-	long x, y;
-	long axisX, axisY, axisZ;
+	float x, y;
+	float axisX, axisY, axisZ;
 	int left,right,center;
 	void updata() {
 		if (GetMouseLeftPress())
